Implement raise method to process AddItemEvent

+++ src/main/java/com/thoughtworks/ddd_workshop/domain/Item.java
@@ -16,7 +16,7 @@ public class Item {
         return name;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return this.quantity;
     }
